--- about.html
@@ -1,51 +1,0 @@
-<!DOCTYPE html>
-
-<html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <title>GAME OF LIFE 2.0</title>
-        <link rel="stylesheet" type="text/css" href="css/about.css">
-    </head>
-    <body>
-        <div id="background">
-            <div id="page">
-                <div id="description">
-		<p>
-		</p>
-                <strong>Game of Life 2.0</strong>
-                <p>This is a version of the classic Game of Life, except
-                with a few changes.  Instead of two states, each cell
-                has 60.  This game was made by a team of 9 programmers
-                and artists in a 24 hour programming tournament held 
-                by Urban Airship in Portland, Oregon.</p>
-                </div>
-                <div id="devs">
-		<p>
-		</p>
-                <strong>Developers:</strong>
-		<p>
-		</p>
-	
-
-                
-                    <li>Aaron Celeste,</li> 
-                    <li>Cezary Wojcik,</li> 
-                    <li>Chance Zibolski, </li>
-                    <li>Damian Kulp, </li>
-                    <li>Fauzi Kliman, </li>
-                    <li>Gabriel Kulp, </li>
-                    <li>Hamzah Khan, </li>
-                    <li>Isaac Archer, </li>
-                    <li>Towaki Takikawa</li>
-                
-		<a href="http://portland.codeday.org/" target="_blank">
- 		<img src="./img/codeday.png" />
-		</a>
-                </div>
-		<p class="center">
-                <a href="splash.html">Home</a>
-		</p>
-            </div>
-        </div>
-    </body>
-</html>
