+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "gameoflife",
+  "version": "3.0.0",
+  "description": "Game of Life 3.0",
+  "main": "server.js",
+  "scripts": {
+    "start": "node server.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Merfoo/GameOfLife-3.0.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Merfoo/GameOfLife-3.0/issues"
+  },
+  "homepage": "https://github.com/Merfoo/GameOfLife-3.0#readme",
+  "type": "module",
+  "dependencies": {
+    "express": "^4.18.2",
+    "path": "^0.12.7"
+  }
+}
