+++ public/css/about.css
@@ -1,0 +1,43 @@
+body {
+    background-color: #ffffff; 
+  background-image: url('../img/Crystal Clouds.png');
+  background-repeat: repeat;
+  background-attachment: fixed;
+  background-position: center;
+}
+strong {
+    line-height: 15px;
+    text-align: center;
+}
+#page {
+    font-size: 1.20em;
+}
+#description {
+    margin-left: auto;
+    margin-right: auto; 
+    margin-bottom: 20px;
+    margin-top: 35px;
+    width: 750px;
+    list-style-type: none;
+    line-height: 1em;
+    border: 2px solid black;
+    border-radius: 20px;
+    height: 100%;
+    background-color: #FFFFFF;
+    padding-left: 14px;
+}
+#devs {
+    margin-left: auto;
+    margin-right: auto; 
+    width: 300px;
+    list-style-type: none;
+    line-height: 1em;
+    border: 2px solid black;
+    border-radius: 20px;
+    height: 100%;
+    background-color: #FFFFFF;
+    padding-left: 14px;
+}
+p.center { 
+text-align: center; 
+}
