+++ public/css/splash.css
@@ -1,0 +1,33 @@
+body {
+    font-size: 2.5em;
+    background-color: #ffffff; 
+  background-image: url('../img/background2.png');
+  background-repeat: repeat;
+  background-attachment: fixed;
+  background-position: center;
+}
+#top {
+    padding-bottom: 10px;
+    margin-left: auto;
+    margin-right: auto; 
+    width: 538px;
+    border: 3px solid black;
+    border-radius: 20px;
+    height: 100%;
+    background-color: #fff;
+}
+#content {
+    margin-left: auto;
+    margin-right: auto; 
+    width: 350px;
+    list-style-type: none;
+    line-height: 1.45em;
+    border: 3px solid black;
+    border-radius: 20px;
+    height: 100%;
+    background-color: #fff;
+}
+
+p.center { 
+text-align: center; 
+}
