+++ public/index.html
@@ -1,0 +1,45 @@
+<!doctype html>
+<html>
+<head>
+  <meta charset="utf-8">
+  <title>
+    GAME OF LIFE 3.0
+  </title>
+
+  <link rel="stylesheet" type="text/css" href="css/base.css">
+  <script src="js/jquery.min.js"></script>
+  <script src="js/util.js"></script>
+  <script src="js/game.js"></script>
+<link rel="icon" href="img/icon.ico" type="image/ico"/>
+</head>
+<body onload="gameInit();">
+<center>
+<div id="logodiv">
+  <canvas id='minimap' width='200' height='200'></canvas>
+  <img id="logo" src="img/Logo.png" alt="Logo" width="538" height="276" />
+</div>
+</center>
+<center>
+<div class="left"></div>
+<div class="gameArea">
+  <canvas id="gameCanvas" width="896" height="640"></canvas>
+  <div id="gameText">test</div>
+<center>
+    <form>
+        <input type="button" id="makegraph" value="Make a Graph"> &nbsp;&nbsp;&nbsp;
+        <input type="button" id="restart" value="Restart"><br />
+        <label for="rSlide">Red</label>
+        <input id="rSlide" type ="range" min ="0" max="255" step ="1" value ="255" /><br />
+        <label for="gSlide">Green</label>
+        <input id="gSlide" type ="range" min ="0" max="255" step ="1" value ="165" /><br />
+        <label for="bSlide">Blue</label>
+        <input id="bSlide" type ="range" min ="0" max="255" step ="1" value ="0" /><br />
+    </form>
+</center><br />
+</div>
+</center>
+<p class="center">
+<a href="splash.html">HOME</a>
+</p>
+</body>
+</html>
