--- splash.html
@@ -1,31 +1,0 @@
-<!doctype html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <title>
-            GAME OF LIFE 2.0
-        </title>
-
-        <link rel="stylesheet" type="text/css" href="css/splash.css">
-        <link rel="icon" href="img/icon.ico" type="image/ico"/>
-    </head>
-    <body>
-
-        <div id="background">
-            <div id="top">
-                <img id="logo" src="img/Logo.png" alt="Logo" width="538" height="276" />
-            </div>
-            <ul id="content">
-	<p>
-	</p>
-                <li><a href="index.html"><img src="img/button.png" alt="start game"></a></li>
-                <li><a href="about.html">ABOUT US</a></li>
-                <li><a href="http://codeday.org">CODE DAY</a></li>
-	<p>
-	</p>
-
-            </ul>
-            </div>
-        </div>
-    </body>
-</html>
